refactor(lib/core): move the base-level stuff (math, wave, dom/browser re-exports, etc) into a core
