+++ src/lib/core/dom.test.ts
