+++ src/lib/core/math.test.ts
