+++ src/lib/core/math.ts
