+++ src/lib/core/util.ts
